added verbose msg generation
