+++ commands/token/token_integration_test.go
@@ -4,7 +4,6 @@
 package token_test
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"os"
@@ -66,13 +65,11 @@ func setup() error {
 
 func testCreateToken(t *testing.T) {
 	var err error
-	var buff bytes.Buffer
 
 	tokFirstCreationTime = time.Now().In(time.UTC)
-	//viper.Set(constants.ArgQuiet, true)
+	viper.Set(constants.ArgQuiet, true)
 
 	c := token.TokenPostCmd()
-	c.Command.SetOut(&buff)
 	err = c.Command.Execute()
 	assert.NoError(t, err)
 
@@ -108,8 +105,8 @@ func testCreateToken(t *testing.T) {
 
 	testToken = *foundTokenViaSdk
 
-	//viper.Reset()
-	//viper.Set(constants.ArgOutput, "text")
+	viper.Reset()
+	viper.Set(constants.ArgOutput, "text")
 }
 
 func testListTokens(t *testing.T) {
@@ -138,10 +135,8 @@ func testGetTokens(t *testing.T) {
 
 func testParseToken(t *testing.T) {
 	var err error
-	var buff bytes.Buffer
 
 	c := token.TokenParseCmd()
-	c.Command.SetOut(&buff) // NOTE: so basically in every command that uses c.Printer.Print(...), overriding just like this isn't enough
 	c.Command.Flags().Set(authservice.ArgToken, *tokenContent.Token)
 
 	err = c.Command.Execute()
